+++ src/main/java/com/zufang/service/UserService.java
@@ -8,13 +8,13 @@ public interface UserService {
 
     public List<User> getUserList(Integer startRow,Integer PageSize);
 
-    public User getUserById(long id);
+    public User getUserById(Integer id);
 
     public int UpdateUserByUser(User user);
 
     public int InsertUser(User user);
 
-    public int deleteUserById(long userId);
+    public int deleteUserById(Integer userId);
 
     public int countUser();
 }
